+++ src/Renderer/Event/AppFragmentEvent.ts
@@ -1,21 +1,11 @@
 import {Event} from './Event';
 
 enum EventNames {
-  showPref = 'showPref',
   showConfigSetup = 'showConfigSetup',
 }
 
-class _PrefEvent {
+class _AppFragmentEvent {
   private readonly event = new Event();
-
-  // show pref
-  emitShowPref() {
-    this.event.emit(EventNames.showPref);
-  }
-
-  onShowPref(owner, handler: () => void) {
-    this.event.on(EventNames.showPref, owner, handler);
-  }
 
   // show config setup
   emitShowConfigSetup() {
@@ -27,4 +17,4 @@ class _PrefEvent {
   }
 }
 
-export const FragmentEvent = new _PrefEvent();
+export const AppFragmentEvent = new _AppFragmentEvent();