+++ src/Renderer/Event/Event.ts
@@ -1,7 +1,9 @@
+import React from 'react';
+
 type HandlerData = {
   eventName: string;
   handler: (...args) => void;
-  owner: any;
+  owner: React.Component;
 }
 
 export class Event {
@@ -18,9 +20,9 @@ export class Event {
     this.handler.push({eventName, owner, handler});
   }
 
-  off(eventName: string, owner: HandlerData['owner']) {
-    this.handler = this.handler.filter(v => v.eventName !== eventName || v.owner !== owner);
-  }
+  // off(eventName: string, owner: HandlerData['owner']) {
+  //   this.handler = this.handler.filter(v => v.eventName !== eventName || v.owner !== owner);
+  // }
 
   offAll(owner: HandlerData['owner']) {
     this.handler = this.handler.filter(v => v.owner !== owner);