+++ src/Renderer/Fragment/Stream/SystemStreamsFragment.tsx
@@ -22,74 +22,34 @@ interface State {
 
 export class SystemStreamsFragment extends React.Component<any, State> {
   state: State = {streams: [], selectedStream: null};
-  private readonly _systemStreamListenerIds: number[] = [];
-  private readonly _streamListenerIds: number[] = [];
-  private readonly _libraryStreamListenerIds: number[] = [];
-  private readonly _issueListenerIds: number[] = [];
 
   componentDidMount() {
     this._loadStreams();
 
-    {
-      let id;
-      id = LibraryStreamEvent.addSelectStreamListener(()=>{
-        this.setState({selectedStream: null});
-      });
-      this._libraryStreamListenerIds.push(id);
-    }
+    LibraryStreamEvent.onSelectStream(this, () => this.setState({selectedStream: null}));
 
-    {
-      let id;
-      id = SystemStreamEvent.addUpdateStreamListener(this._loadStreams.bind(this));
-      this._systemStreamListenerIds.push(id);
+    SystemStreamEvent.onUpdateStream(this, this._loadStreams.bind(this));
+    SystemStreamEvent.onSelectStream(this, (stream)=>{
+      if (stream.enabled) this.setState({selectedStream: stream});
+    });
+    SystemStreamEvent.onRestartAllStreams(this, this._loadStreams.bind(this));
 
-      id = SystemStreamEvent.addSelectStreamListener((stream)=>{
-        if (stream.enabled) this.setState({selectedStream: stream});
-      });
-      this._systemStreamListenerIds.push(id);
+   StreamEvent.onUpdateStream(this, this._loadStreams.bind(this));
+   StreamEvent.onSelectStream(this, () => this.setState({selectedStream: null}));
+   StreamEvent.onRestartAllStreams(this, this._loadStreams.bind(this));
 
-      id = SystemStreamEvent.addRestartAllStreamsListener(this._loadStreams.bind(this));
-      this._systemStreamListenerIds.push(id);
-    }
-
-    {
-      let id;
-      id = StreamEvent.addUpdateStreamListener(this._loadStreams.bind(this));
-      this._streamListenerIds.push(id);
-
-      id = StreamEvent.addSelectStreamListener(()=>{
-        this.setState({selectedStream: null});
-      });
-      this._streamListenerIds.push(id);
-
-      id = StreamEvent.addRestartAllStreamsListener(this._loadStreams.bind(this));
-      this._streamListenerIds.push(id);
-    }
-
-    {
-      let id;
-      id = IssueEvent.addReadIssueListener(this._loadStreams.bind(this));
-      this._issueListenerIds.push(id);
-
-      id = IssueEvent.addReadIssuesListener(this._loadStreams.bind(this));
-      this._issueListenerIds.push(id);
-
-      id = IssueEvent.addArchiveIssueListener(this._loadStreams.bind(this));
-      this._issueListenerIds.push(id);
-
-      id = IssueEvent.addReadAllIssuesListener(this._loadStreams.bind(this));
-      this._issueListenerIds.push(id);
-
-      id = IssueEvent.addReadAllIssuesFromLibraryListener(this._loadStreams.bind(this));
-      this._issueListenerIds.push(id);
-    }
+    IssueEvent.onReadIssue(this, this._loadStreams.bind(this));
+    IssueEvent.onReadIssues(this, this._loadStreams.bind(this));
+    IssueEvent.addArchiveIssueListener(this, this._loadStreams.bind(this));
+    IssueEvent.onReadAllIssues(this, this._loadStreams.bind(this));
+    IssueEvent.onReadAllIssuesFromLibrary(this, this._loadStreams.bind(this));
   }
 
   componentWillUnmount() {
-    SystemStreamEvent.removeListeners(this._systemStreamListenerIds);
-    StreamEvent.removeListeners(this._streamListenerIds);
-    LibraryStreamEvent.removeListeners(this._libraryStreamListenerIds);
-    IssueEvent.removeListeners(this._issueListenerIds);
+    SystemStreamEvent.offAll(this);
+    StreamEvent.offAll(this);
+    LibraryStreamEvent.offAll(this);
+    IssueEvent.offAll(this);
   }
 
   async _loadStreams() {