+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+- **Break**
+  - Remove import CSS theme feature.
+- **Internal**
+  - Change architecture what use stream in renderer process. [#146](https://github.com/jasperapp/jasper/pull/146)
+
 # 0.9.3 (2020/07/12)
 - **Feat**
   - Support to automatically load cookie from BrowserWindow. [#140](https://github.com/jasperapp/jasper/pull/140)